dismantled torch_rl_algorithm
